requirements.txt
asgiref==3.8.1
Django==5.1.1 
sqlparse==0.5.1
tzdata==2024.2
python-dotenv
gunicorn
whitenoise
openai-whisper
elevenlabs
openai
langchain
unstructured
langchain_openai
chromadb
pymupdf